--- a/Project.toml
+++ b/Project.toml
@@ -16,7 +16,7 @@ QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
-Compat = "3"
+Compat = "3, 4"
 HDF5 = "0.15, 0.16"
 ITensors = "0.3"
 Infinities = "0.1"