This includes a few optimizations:

1. Use `eager=true` in `eigsolve`, which tries to stop before building the entire Krylov space if the precision is met (https://jutho.github.io/KrylovKit.jl/stable/man/algorithms/#KrylovKit.Lanczos).
2. Improve contraction sequences in computing AC and C in the `InfiniteSum{MPO}` version of VUMPS. Notably, the previous version contracted each MPO term from left to right, whereas it is more efficient to contract the MPO towards the position of AC and C. There are more optimizations to do along those lines beyond what is done here.

Results from this PR for the extended Ising model example (`vumps_ising_extended.jl`):
```julia
maxdim = 64 # Maximum bond dimension
cutoff = 1e-6 # Singular value cutoff when increasing the bond dimension
max_vumps_iters = 100 # Maximum number of iterations of the VUMPS algorithm at each bond dimension
vumps_tol = 1e-6
outer_iters = 10 # Number of times to increase the bond dimension
eager = true # Introduced in this PR

N = 3# Number of sites in the unit cell
J = -1.0
J₂ = -0.2
h = 1.0;
```
```julia
Run VUMPS with new bond dimension
Using VUMPS solver with time step -Inf
Running VUMPS with multisite_update_alg = sequential
VUMPS iteration 1 (out of maximum 100). Bond dimension = 64, energy = ([-1.250843628949263, -1.2508436289529703, -1.250843628961579], [-1.25084362894743, -1.250843628991484, -1.2508436289599254]), ϵᵖʳᵉˢ = 1.3059754452205646e-9, tol = 1.0e-6
Precision error 1.3059754452205646e-9 reached tolerance 1.0e-6, stopping VUMPS after 1 iterations (of a maximum 100).
  0.221690 seconds (2.41 M allocations: 399.215 MiB, 12.88% gc time)
  5.815601 seconds (61.68 M allocations: 8.900 GiB, 16.68% gc time)
```
and on the same run on the `main` branch:
```julia
Run VUMPS with new bond dimension
Using VUMPS solver with time step -Inf
Running VUMPS with multisite_update_alg = sequential
VUMPS iteration 1 (out of maximum 100). Bond dimension = 64, energy = ([-1.250843628952611, -1.2508436289526905, -1.2508436289441933], [-1.2508436289520395, -1.2508436289507119, -1.2508436289511176]), ϵᵖʳᵉˢ = 4.736627585934926e-9, tol = 1.0e-6
Precision error 4.736627585934926e-9 reached tolerance 1.0e-6, stopping VUMPS after 1 iterations (of a maximum 100).
 41.533461 seconds (5.11 M allocations: 93.863 GiB, 20.41% gc time)
152.027322 seconds (141.63 M allocations: 365.922 GiB, 19.16% gc time, 0.00% compilation time)
```

For completeness, this is the current branch with `eager=false` (so the same `eigsolve` settings as before):
```julia
Run VUMPS with new bond dimension
Using VUMPS solver with time step -Inf
Running VUMPS with multisite_update_alg = sequential
VUMPS iteration 1 (out of maximum 100). Bond dimension = 56, energy = ([-1.2508436289529472, -1.250843628952671, -1.2508436289468248], [-1.2508436289536915, -1.2508436289563363, -1.2508436289524103]), ϵᵖʳᵉˢ = 4.736836655850174e-9, tol = 1.0e-6
Precision error 4.736836655850174e-9 reached tolerance 1.0e-6, stopping VUMPS after 1 iterations (of a maximum 100).
  0.445129 seconds (4.40 M allocations: 750.770 MiB, 16.77% gc time)
  9.479867 seconds (113.59 M allocations: 15.385 GiB, 19.08% gc time)
```
So `eager=true` indeed speed things up without losing accuracy.

Another change I made was making sure the phases of `AC` and `C` match when computing the precision error. Sometimes they were mismatched so the precision error (`|AC - AL * C|`) was close to `2` instead of close to `0` (in the case of real tensors when the phase degree of freedom of the eigenvectors `AC` and `C` are `+/-1`). That could potentially cause convergence issues, since the precision of `eigsolve` is determined automatically from the precision error.

@LHerviou @hershsingh